Split server out to separate crate
