+++ crates/brace-cms-server/src/lib.rs
@@ -3,21 +3,11 @@ use std::net::Ipv4Addr;
 
 use brace_config::Config;
 use brace_web::core::middleware::Logger;
-use brace_web::core::{web, App, HttpResponse, HttpServer};
+use brace_web::core::{web, App, HttpServer};
 
-pub mod store {
-    pub mod postgres {
-        pub use brace_cms_store_postgres::*;
-    }
-}
-
-async fn index() -> HttpResponse {
-    HttpResponse::Ok().body("Hello world")
-}
+pub mod routes;
 
 pub async fn server(config: Config) -> io::Result<()> {
-    brace_cms_log::init(&config).unwrap();
-
     let host = config
         .get::<_, Ipv4Addr>("server.host")
         .unwrap_or_else(|_| Ipv4Addr::new(127, 0, 0, 1));
@@ -31,7 +21,7 @@ pub async fn server(config: Config) -> io::Result<()> {
         App::new()
             .wrap(Logger::new(format))
             .configure(postgres.clone())
-            .route("/", web::get().to(index))
+            .route("/", web::get().to(crate::routes::index::get))
     });
 
     #[cfg(feature = "dev")]
@@ -52,21 +42,4 @@ pub async fn server(config: Config) -> io::Result<()> {
     }
 
     server.run().await
-}
-
-#[cfg(test)]
-mod tests {
-    use brace_web::core::test::{call_service, init_service, TestRequest};
-    use brace_web::core::{web, App};
-
-    use super::index;
-
-    #[actix_rt::test]
-    async fn test_server_index_get() {
-        let mut app = init_service(App::new().route("/", web::get().to(index))).await;
-        let req = TestRequest::with_header("content-type", "text/plain").to_request();
-        let res = call_service(&mut app, req).await;
-
-        assert!(res.status().is_success());
-    }
 }
